Fix all Run #7 compile errors
- Add GlobalUsings.cs: 'global using System.IO' (not in Worker SDK implicit usings)
  and 'global using ActivityEvent = WinDiagSvc.Models.ActivityEvent' to resolve
  ambiguity with System.Diagnostics.ActivityEvent (.NET 8 tracing type).
  Fixes ~50 Path/File/Directory/FileInfo/SearchOption/Stream errors across all files.

- Fix CS1626 in EventLogWatcherService.DiscoverActiveChannels:
  moved 'yield return' outside try/catch (C# language restriction).

- Fix CS0266 in ScreenshotWorker.CaptureActiveWindow:
  added explicit (uint) cast for biWidth field (BITMAPINFOHEADER).

- Wire OnWindowChanged and OnUiEvent screenshot triggers in Program.cs:
  registered WindowWatcher/ScreenshotWorker/UiAutomationCapture as singletons
  so they can be resolved after Build(), then connected the delegates.
  Fixes CS0649 warnings and makes event-driven screenshots actually work.

+++ src/WinDiagSvc/Capture/ScreenshotWorker.cs
@@ -124,8 +124,8 @@ public sealed class ScreenshotWorker : BackgroundService
             var bmpInfo = new BITMAPINFOHEADER
             {
                 biSize        = (uint)Marshal.SizeOf<BITMAPINFOHEADER>(),
-                biWidth       = w,
-                biHeight      = -h,   // top-down
+                biWidth       = (uint)w,   // biWidth is uint; w is int
+                biHeight      = -h,       // top-down DIB (negative height)
                 biPlanes      = 1,
                 biBitCount    = 32,
                 biCompression = BI_RGB,
